+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Changes
 
 - Added additional framework target `net9.0`
-- `ResolveWithService` and `ResolveWithServices` methods are now marked obsolete. Please just use `Resolve` in the same way
+- `ResolveWithService` and `ResolveWithServices` methods are now marked obsolete. Please use `Resolve` in the same way
+- #430 - `MapGraphQL` has new parameter `followSpec` which changes the behavior to follow https://github.com/graphql/graphql-over-http/blob/main/spec/GraphQLOverHTTP.md and is `false` by default in v5.x. If set to `true` it will follow the spec, the major change being that if the HTTP request was valid you'll get a `200` status code even if there may be GraphQL errors in the response. _Note this will be the default behavior in version v6._
 
 ## Fixes
 
@@ -22,7 +23,7 @@ schema.Type<User>().AddField("username", u => u.Name, "Username")
 
 ## Fixes
 
-- #432 - filter expression was incorrectly using bit-wise OR or OrElse
+- #432 - filter expression was incorrectly using bit-wise `Or` instead of `OrElse`
 
 # 5.5.2
 
@@ -382,7 +383,7 @@ schema.ReplaceField("people",
 
 ## Breaking Changes
 
-- `EntityGraphQL.AspNet` now targets `net6.0` and `net7.0`, dropping tagets `netcoreapp3.1` or `net5.0`. You can still use the base `EntityGraphQL` library with older targets.
+- `EntityGraphQL.AspNet` now targets `net6.0` and `net7.0`, dropping targets `netcoreapp3.1` and `net5.0`. You can still use the base `EntityGraphQL` library with older targets.
 - Interface `IExposableException` has been removed. Use `SchemaBuilderSchemaOptions.AllowedExceptions` or the new `AllowedExceptionAttribute` to define which exceptions are rendered into the results
 - #254 - Previously passing `null` for the `ClaimsPrincipal` in `ExecuteRequest()` would skip any authorization checks. All authorization checks are now done regardless of the `ClaimsPrincipal` value. Meaning `null` will fail if there is fields requiring authorization.
 - `IDirectiveProcessor` interface has changed. See upgrade docs for changes